unittests added

+++ PetrolEngine.cpp
@@ -7,7 +7,7 @@
 constexpr bool NEUTRAL_GEAR = 0;
 constexpr short REAR_GEAR = -1;
 
-PetrolEngine::PetrolEngine(uint16_t power, size_t capacity, uint8_t gears)
+PetrolEngine::PetrolEngine(uint16_t power, uint16_t capacity, uint8_t gears)
     : Engine(power), capacity_(capacity), gears_(gears), currentGear_(0) {
     std::cout << __FUNCTION__ << std::endl;
 }
